refactor(contracts): drop v0.7 path entirely — projects-jsonl fold is the only model
