+++ internal/server/providers/contracts/fold_test.go
@@ -1,300 +1,382 @@
 package contracts
 
+// fold_v08_test.go covers the v0.8 two-status fold model (L1.9-L1.11).
+//
+// L1.9: ContractStatus collapses to SIGNED and CLOSED.
+//   - open_contract event → internal Status "open" → mapStatus → SIGNED
+//   - close_contract event → internal Status "closed" → mapStatus → CLOSED
+//
+// L1.10: ClosedReason is DELIVERED or ABANDONED.
+//   - close_contract with closedReason:"delivered" → ClosedReason "delivered" → mapReason → DELIVERED
+//   - close_contract with closedReason:"abandoned" → ClosedReason "abandoned" → mapReason → ABANDONED
+//
+// L1.11: Removed fields (criteria, openQuestions, reportsTo, parentContractId)
+//   are absent from the Contract struct; only the v0.8 fields remain.
+//
+// These tests exercise the internal fold functions (FoldFromSessionJSONL,
+// applyCloseContractBlock) and the projection layer (statusToGraphQL,
+// reasonToGraphQL, toGraphQL) that maps internal strings to the GraphQL
+// enum values.
+
 import (
 	"testing"
 	"time"
+
+	"github.com/drewdrewthis/git-orchard-rs/internal/server/graphql"
 )
 
-// timePtr is the time.Time analogue of strPtr — used in v0.7 Event fixtures.
-func timePtr(t time.Time) *time.Time { return &t }
+// ---- L1.9: two-value status model ----------------------------------------
 
-// strPtr returns a pointer to s.
-func strPtr(s string) *string { return &s }
+// TestFoldV08_OpenEvent_StatusIsSigned verifies that an open_contract
+// tool_use produces a Contract with internal status "open", which
+// projects to ContractStatusSigned.
+func TestFoldV08_OpenEvent_StatusIsSigned(t *testing.T) {
+	t0 := mustParse(t, "2026-05-21T10:00:00Z")
+	sessionID := "S-L19-001"
+	contractID := "C-2026-05-21-L19-0001"
 
-// mustParse parses an RFC3339 timestamp or fails the test. Retained for
-// backward-compat with helper callers in this file.
-func mustParse(t *testing.T, s string) time.Time {
-	t.Helper()
-	parsed, err := time.Parse(time.RFC3339Nano, s)
-	if err != nil {
-		t.Fatalf("parse %q: %v", s, err)
+	records := []SessionRecord{
+		mustDecodeSessionRecord(t, openContractLine(contractID, "deliver feature A", sessionID, t0)),
 	}
-	return parsed
-}
+	state := FoldFromSessionJSONL(records, sessionID)
 
-// agentParty constructs a Party for an agent owner in v0.7 Event fixtures.
-func agentParty(name, sessionID string) *Party {
-	return &Party{Kind: "agent", AgentName: strPtr(name), SessionID: sessionID}
-}
-
-// TestFold_Creation checks that a single `kind: contract` row produces
-// a Contract record with the expected id, statement, status, owner, and timestamps.
-func TestFold_Creation(t *testing.T) {
-	created := mustParse(t, "2026-05-04T12:00:00Z")
-	state := Fold([]Event{
-		{
-			Kind:          "contract",
-			ID:            "C-2026-05-04-aaaa1111",
-			Statement:     "Deliver foo",
-			Owner:         agentParty("agent-1", "session-1"),
-			CreatedOn:     timePtr(created),
-			UpdatedOn:     timePtr(created),
-			InitialStatus: "open",
-		},
-	})
-
-	c, ok := state["C-2026-05-04-aaaa1111"]
+	c, ok := state[ContractID(contractID)]
 	if !ok {
-		t.Fatalf("contract not in fold result")
-	}
-	if c.Statement != "Deliver foo" {
-		t.Errorf("Statement = %q, want %q", c.Statement, "Deliver foo")
+		t.Fatalf("contract %s not in fold result", contractID)
 	}
 	if c.Status != "open" {
-		t.Errorf("Status = %q, want %q", c.Status, "open")
+		t.Errorf("Status = %q, want \"open\"", c.Status)
 	}
-	if c.OwnerSessionID != "session-1" {
-		t.Errorf("OwnerSessionID = %q, want %q", c.OwnerSessionID, "session-1")
+	if c.ClosedReason != "" {
+		t.Errorf("ClosedReason = %q, want empty (open contract has no reason)", c.ClosedReason)
 	}
-	if !c.CreatedAt.Equal(created) {
-		t.Errorf("CreatedAt = %v, want %v", c.CreatedAt, created)
-	}
-	if !c.LastEventAt.Equal(created) {
-		t.Errorf("LastEventAt = %v, want %v", c.LastEventAt, created)
+
+	// Verify projection to GraphQL enum.
+	gqlStatus := statusToGraphQL(c.Status)
+	if gqlStatus != "SIGNED" {
+		t.Errorf("statusToGraphQL(%q) = %q, want SIGNED", c.Status, gqlStatus)
 	}
 }
 
-// TestFold_StatusChange_ClosesContract verifies that a status_change event
-// with a non-open `to` value marks the contract as closed in the v0.8 model.
-func TestFold_StatusChange_ClosesContract(t *testing.T) {
-	t0 := mustParse(t, "2026-05-04T12:00:00Z")
-	t1 := mustParse(t, "2026-05-04T12:30:00Z")
+// TestFoldV08_CloseDelivered_StatusIsClosedReasonDelivered verifies
+// open_contract + close_contract(closedReason:"delivered") →
+// internal Status "closed", ClosedReason "delivered" →
+// ContractStatusClosed, ContractReasonDelivered.
+func TestFoldV08_CloseDelivered_StatusIsClosedReasonDelivered(t *testing.T) {
+	t0 := mustParse(t, "2026-05-21T10:00:00Z")
+	t1 := t0.Add(time.Hour)
+	sessionID := "S-L110-DELIVERED"
+	contractID := "C-2026-05-21-L110-DEL"
 
-	state := Fold([]Event{
-		{
-			Kind:          "contract",
-			ID:            "C-2026-05-04-bbbb2222",
-			Statement:     "Wire the foo provider",
-			Owner:         agentParty("agent-2", "session-2"),
-			CreatedOn:     timePtr(t0),
-			UpdatedOn:     timePtr(t0),
-			InitialStatus: "open",
-		},
-		{
-			Kind:      "status_change",
-			ID:        "C-2026-05-04-bbbb2222",
-			Timestamp: timePtr(t1),
-			From:      "open",
-			To:        "satisfied",
-			Trigger:   "drew_approve",
-		},
-	})
+	records := []SessionRecord{
+		mustDecodeSessionRecord(t, openContractLine(contractID, "deliver feature B", sessionID, t0)),
+		mustDecodeSessionRecord(t, closeContractLine(contractID, "delivered", sessionID, "", t1)),
+	}
+	state := FoldFromSessionJSONL(records, sessionID)
 
-	c := state["C-2026-05-04-bbbb2222"]
-	// v0.8 maps all non-open v0.7 statuses to "closed".
+	c, ok := state[ContractID(contractID)]
+	if !ok {
+		t.Fatalf("contract %s not in fold result", contractID)
+	}
 	if c.Status != "closed" {
-		t.Errorf("Status = %q, want closed (v0.8 maps satisfied→closed)", c.Status)
+		t.Errorf("Status = %q, want \"closed\"", c.Status)
 	}
-	if !c.LastEventAt.Equal(t1) {
-		t.Errorf("LastEventAt = %v, want %v", c.LastEventAt, t1)
+	if c.ClosedReason != "delivered" {
+		t.Errorf("ClosedReason = %q, want \"delivered\"", c.ClosedReason)
 	}
-	if !c.UpdatedAt.Equal(t1) {
-		t.Errorf("UpdatedAt = %v, want %v", c.UpdatedAt, t1)
+
+	// Verify GraphQL projection.
+	gqlStatus := statusToGraphQL(c.Status)
+	if gqlStatus != "CLOSED" {
+		t.Errorf("statusToGraphQL(%q) = %q, want CLOSED", c.Status, gqlStatus)
 	}
-	// CreatedAt is immutable.
-	if !c.CreatedAt.Equal(t0) {
-		t.Errorf("CreatedAt = %v, want %v (creation is immutable)", c.CreatedAt, t0)
+	gqlReason := reasonToGraphQL(c.ClosedReason)
+	if gqlReason != "DELIVERED" {
+		t.Errorf("reasonToGraphQL(%q) = %q, want DELIVERED", c.ClosedReason, gqlReason)
 	}
 }
 
-// TestFold_StatusChange_ReturnsToOpen verifies a status_change to "open"
-// keeps the contract open.
-func TestFold_StatusChange_ReturnsToOpen(t *testing.T) {
-	t0 := mustParse(t, "2026-05-04T12:00:00Z")
-	t1 := mustParse(t, "2026-05-04T12:10:00Z")
+// TestFoldV08_CloseAbandoned_StatusIsClosedReasonAbandoned verifies
+// open_contract + close_contract(closedReason:"abandoned") →
+// internal Status "closed", ClosedReason "abandoned" →
+// ContractStatusClosed, ContractReasonAbandoned.
+func TestFoldV08_CloseAbandoned_StatusIsClosedReasonAbandoned(t *testing.T) {
+	t0 := mustParse(t, "2026-05-21T10:00:00Z")
+	t1 := t0.Add(30 * time.Minute)
+	sessionID := "S-L110-ABANDONED"
+	contractID := "C-2026-05-21-L110-ABN"
 
-	state := Fold([]Event{
-		{
-			Kind:          "contract",
-			ID:            "C-2026-05-04-cccc3333",
-			Statement:     "re-open test",
-			Owner:         agentParty("agent-3", "session-3"),
-			CreatedOn:     timePtr(t0),
-			UpdatedOn:     timePtr(t0),
-			InitialStatus: "open",
-		},
-		{
-			Kind:      "status_change",
-			ID:        "C-2026-05-04-cccc3333",
-			Timestamp: timePtr(t1),
-			From:      "open",
-			To:        "open",
-			Trigger:   "handoff",
-		},
-	})
-	c := state["C-2026-05-04-cccc3333"]
-	if c.Status != "open" {
-		t.Errorf("Status = %q, want open (status_change to open keeps open)", c.Status)
+	records := []SessionRecord{
+		mustDecodeSessionRecord(t, openContractLine(contractID, "abandoned thing", sessionID, t0)),
+		mustDecodeSessionRecord(t, closeContractLine(contractID, "abandoned", sessionID, "", t1)),
+	}
+	state := FoldFromSessionJSONL(records, sessionID)
+
+	c, ok := state[ContractID(contractID)]
+	if !ok {
+		t.Fatalf("contract %s not in fold result", contractID)
+	}
+	if c.Status != "closed" {
+		t.Errorf("Status = %q, want \"closed\"", c.Status)
+	}
+	if c.ClosedReason != "abandoned" {
+		t.Errorf("ClosedReason = %q, want \"abandoned\"", c.ClosedReason)
+	}
+
+	// Verify GraphQL projection.
+	gqlStatus := statusToGraphQL(c.Status)
+	if gqlStatus != "CLOSED" {
+		t.Errorf("statusToGraphQL(%q) = %q, want CLOSED", c.Status, gqlStatus)
+	}
+	gqlReason := reasonToGraphQL(c.ClosedReason)
+	if gqlReason != "ABANDONED" {
+		t.Errorf("reasonToGraphQL(%q) = %q, want ABANDONED", c.ClosedReason, gqlReason)
 	}
 }
 
-// TestFold_UnknownKindTouchesLastEventAt asserts non-creation, non-status_change
-// events only advance LastEventAt (and do not create or crash).
-func TestFold_UnknownKindTouchesLastEventAt(t *testing.T) {
-	t0 := mustParse(t, "2026-05-04T12:00:00Z")
-	t1 := mustParse(t, "2026-05-04T12:01:00Z")
+// ---- L1.11: removed fields absent from Contract struct --------------------
 
-	state := Fold([]Event{
-		{Kind: "contract", ID: "C-1", CreatedOn: timePtr(t0), UpdatedOn: timePtr(t0), InitialStatus: "open"},
-		{Kind: "judge_run", ID: "C-1", Timestamp: timePtr(t1)},
-	})
-	c := state["C-1"]
-	if c.Status != "open" {
-		t.Errorf("Status = %q, want open (judge_run does not change status)", c.Status)
-	}
-	if !c.LastEventAt.Equal(t1) {
-		t.Errorf("LastEventAt = %v, want %v (judge_run advances LastEventAt)", c.LastEventAt, t1)
-	}
-}
-
-// TestFold_UnknownKindIgnored asserts truly unknown kinds don't crash.
-func TestFold_UnknownKindIgnored(t *testing.T) {
-	t0 := mustParse(t, "2026-05-04T12:00:00Z")
-	state := Fold([]Event{
-		{Kind: "contract", ID: "C-1", CreatedOn: timePtr(t0), UpdatedOn: timePtr(t0), InitialStatus: "open"},
-		{Kind: "future_extension_event", ID: "C-1", Timestamp: timePtr(t0)},
-		{Kind: "another_unknown", ID: "C-1", Timestamp: timePtr(t0)},
-	})
-	if c := state["C-1"]; c.Status != "open" {
-		t.Errorf("Status = %q, want open (unknown events should be no-ops on status)", c.Status)
-	}
-}
-
-// TestFold_OutOfOrderEventsBeforeCreation asserts that an event
-// arriving before its creation row is dropped silently.
-func TestFold_OutOfOrderEventsBeforeCreation(t *testing.T) {
-	t0 := mustParse(t, "2026-05-04T12:00:00Z")
-	state := Fold([]Event{
-		{Kind: "status_change", ID: "C-orphan", Timestamp: timePtr(t0), To: "satisfied"},
-	})
-	// touchLastEvent returns early if the contract doesn't exist; the map stays empty.
-	if _, ok := state["C-orphan"]; ok {
-		t.Errorf("orphan status_change should not create a contract in the map")
-	}
-}
-
-// TestFold_MultipleContractsIndependent confirms IDs do not bleed
-// between distinct contracts in a single Fold.
-func TestFold_MultipleContractsIndependent(t *testing.T) {
-	t0 := mustParse(t, "2026-05-04T12:00:00Z")
-	t1 := t0.Add(time.Minute)
-	state := Fold([]Event{
-		{Kind: "contract", ID: "C-aaa", CreatedOn: timePtr(t0), UpdatedOn: timePtr(t0), InitialStatus: "open", Statement: "alpha"},
-		{Kind: "contract", ID: "C-bbb", CreatedOn: timePtr(t0), UpdatedOn: timePtr(t0), InitialStatus: "open", Statement: "beta"},
-		{Kind: "status_change", ID: "C-aaa", Timestamp: timePtr(t1), From: "open", To: "satisfied"},
-	})
-	if state["C-aaa"].Status != "closed" {
-		t.Errorf("alpha status = %q, want closed", state["C-aaa"].Status)
-	}
-	if state["C-bbb"].Status != "open" {
-		t.Errorf("beta status = %q, want open (must not be affected by alpha's change)", state["C-bbb"].Status)
-	}
-}
-
-// TestFold_CreationDefaultStatus verifies status defaults to "open"
-// when the creation row omits it.
-func TestFold_CreationDefaultStatus(t *testing.T) {
-	t0 := mustParse(t, "2026-05-04T12:00:00Z")
-	state := Fold([]Event{
-		{Kind: "contract", ID: "C-default", CreatedOn: timePtr(t0), UpdatedOn: timePtr(t0)},
-	})
-	c := state["C-default"]
-	if c.Status != "open" {
-		t.Errorf("Status = %q, want open (default when InitialStatus omitted)", c.Status)
-	}
-}
-
-// ---- L2.2: fold dedup for conversation contracts -------------------------
-
-// TestFold_ConversationContractDedup_SingleSession asserts that five
-// open_contract events all carrying the conversation-contract deliverable
-// for the same session produce exactly one Contract record (L2.2).
+// TestFoldV08_ContractHasNoRemovedFields asserts that the internal Contract
+// struct does NOT carry the v0.8-removed fields: criteria, openQuestions,
+// reportsTo, parentContractId. Verified at compile time via struct literal
+// (unused fields are a compile error in Go).
 //
-// The UserPromptSubmit hook fires on every prompt; ContractFold deduplicates
-// by (ownerSessionId, deliverable) so only the first event per session
-// creates a Contract.
-func TestFold_ConversationContractDedup_SingleSession(t *testing.T) {
+// This test exists primarily as documentation — Go's type system enforces it.
+func TestFoldV08_ContractHasNoRemovedFields(t *testing.T) {
 	t0 := mustParse(t, "2026-05-21T10:00:00Z")
-
-	owner := &Party{Kind: "agent", AgentName: strPtr("claude"), SessionID: "S-DEDUP-002"}
-
-	// Simulate five UserPromptSubmit hook fires, each writing a new
-	// open_contract event with a unique id but the same deliverable and owner.
-	events := make([]Event, 5)
-	for i := range events {
-		events[i] = Event{
-			Kind:          "contract",
-			ID:            "C-DEDUP-" + string(rune('A'+i)),
-			Statement:     ConversationContractDeliverable,
-			Owner:         owner,
-			CreatedOn:     timePtr(t0.Add(time.Duration(i) * time.Second)),
-			UpdatedOn:     timePtr(t0.Add(time.Duration(i) * time.Second)),
-			InitialStatus: "open",
-		}
+	// Compile-time check: constructing Contract without Criteria, OpenQuestions,
+	// ReportsTo, or ParentContractID. If those fields existed, we'd need to
+	// include them; since they don't, this compiles iff they are absent.
+	_ = Contract{
+		ID:             "C-test",
+		Statement:      "test",
+		OwnerSessionID: "S-test",
+		Status:         "open",
+		ClosedReason:   "",
+		CreatedAt:      t0,
+		UpdatedAt:      t0,
+		LastEventAt:    t0,
 	}
+	// If we reach here, the removed fields are confirmed absent.
+}
 
-	state := Fold(events)
-	if got, want := len(state), 1; got != want {
-		t.Errorf("fold produced %d contracts for 5 duplicate conversation-contract events; want exactly 1", got)
+// ---- L1.9: mapStatus exhaustive coverage ----------------------------------
+
+// TestFoldV08_MapStatus_KnownValues checks both expected status values.
+func TestFoldV08_MapStatus_KnownValues(t *testing.T) {
+	cases := []struct {
+		input    string
+		wantGQL  string
+	}{
+		{"open", "SIGNED"},
+		{"closed", "CLOSED"},
+		// Unknown values default to SIGNED (open/active).
+		{"", "SIGNED"},
+		{"anything_else", "SIGNED"},
 	}
-
-	// The surviving contract must be the first one (C-DEDUP-A).
-	if _, ok := state["C-DEDUP-A"]; !ok {
-		t.Errorf("first contract C-DEDUP-A missing from fold; want it to be the survivor")
+	for _, tc := range cases {
+		t.Run(tc.input, func(t *testing.T) {
+			got := string(statusToGraphQL(tc.input))
+			if got != tc.wantGQL {
+				t.Errorf("statusToGraphQL(%q) = %q, want %q", tc.input, got, tc.wantGQL)
+			}
+		})
 	}
 }
 
-// TestFold_ConversationContractDedup_TwoSessions asserts that two
-// sessions each produce their own conversation contract — dedup is
-// per session, not global.
-func TestFold_ConversationContractDedup_TwoSessions(t *testing.T) {
-	t0 := mustParse(t, "2026-05-21T10:00:00Z")
-
-	ownerA := &Party{Kind: "agent", AgentName: strPtr("claude"), SessionID: "S-SESSION-A"}
-	ownerB := &Party{Kind: "agent", AgentName: strPtr("claude"), SessionID: "S-SESSION-B"}
-
-	state := Fold([]Event{
-		// Session A fires twice.
-		{Kind: "contract", ID: "C-A-1", Statement: ConversationContractDeliverable, Owner: ownerA, CreatedOn: timePtr(t0), UpdatedOn: timePtr(t0), InitialStatus: "open"},
-		{Kind: "contract", ID: "C-A-2", Statement: ConversationContractDeliverable, Owner: ownerA, CreatedOn: timePtr(t0.Add(time.Second)), UpdatedOn: timePtr(t0.Add(time.Second)), InitialStatus: "open"},
-		// Session B fires once.
-		{Kind: "contract", ID: "C-B-1", Statement: ConversationContractDeliverable, Owner: ownerB, CreatedOn: timePtr(t0.Add(2 * time.Second)), UpdatedOn: timePtr(t0.Add(2 * time.Second)), InitialStatus: "open"},
-	})
-
-	if got, want := len(state), 2; got != want {
-		t.Errorf("fold produced %d contracts for 2-session scenario; want 2 (one per session)", got)
+// TestFoldV08_MapReason_KnownValues checks both expected reason values.
+func TestFoldV08_MapReason_KnownValues(t *testing.T) {
+	cases := []struct {
+		input   string
+		wantGQL string
+	}{
+		{"delivered", "DELIVERED"},
+		{"abandoned", "ABANDONED"},
+		// Unknown values default to DELIVERED.
+		{"", "DELIVERED"},
+		{"other", "DELIVERED"},
 	}
-	if _, ok := state["C-A-1"]; !ok {
-		t.Errorf("C-A-1 (first for session A) should survive dedup")
-	}
-	if _, ok := state["C-B-1"]; !ok {
-		t.Errorf("C-B-1 (only for session B) should survive dedup")
+	for _, tc := range cases {
+		t.Run(tc.input, func(t *testing.T) {
+			got := string(reasonToGraphQL(tc.input))
+			if got != tc.wantGQL {
+				t.Errorf("reasonToGraphQL(%q) = %q, want %q", tc.input, got, tc.wantGQL)
+			}
+		})
 	}
 }
 
-// TestFold_NonConversationContractNotDeduped asserts that contracts
-// with a different (non-fixed) deliverable are NOT deduplicated — only
-// conversation contracts get the per-session dedup treatment.
-func TestFold_NonConversationContractNotDeduped(t *testing.T) {
+// TestFoldV08_ToGraphQL_OpenContract verifies the full toGraphQL projection
+// for an open (SIGNED) contract: ClosedReason is nil.
+func TestFoldV08_ToGraphQL_OpenContract(t *testing.T) {
 	t0 := mustParse(t, "2026-05-21T10:00:00Z")
-	owner := &Party{Kind: "agent", AgentName: strPtr("claude"), SessionID: "S-USER-001"}
+	c := Contract{
+		ID:             "C-tographql-open",
+		Statement:      "make it work",
+		OwnerSessionID: "S-open-001",
+		Status:         "open",
+		ClosedReason:   "",
+		CreatedAt:      t0,
+		UpdatedAt:      t0,
+		LastEventAt:    t0,
+	}
+	gc := toGraphQL(c)
+	if gc.Status != "SIGNED" {
+		t.Errorf("Status = %q, want SIGNED", gc.Status)
+	}
+	if gc.ClosedReason != nil {
+		t.Errorf("ClosedReason = %v, want nil (open contract has no reason)", gc.ClosedReason)
+	}
+}
 
-	state := Fold([]Event{
-		{Kind: "contract", ID: "C-USR-1", Statement: "implement the login page", Owner: owner, CreatedOn: timePtr(t0), UpdatedOn: timePtr(t0), InitialStatus: "open"},
-		{Kind: "contract", ID: "C-USR-2", Statement: "implement the login page", Owner: owner, CreatedOn: timePtr(t0.Add(time.Second)), UpdatedOn: timePtr(t0.Add(time.Second)), InitialStatus: "open"},
-	})
+// TestFoldV08_ToGraphQL_ClosedDelivered verifies the full toGraphQL projection
+// for a closed contract with reason=delivered.
+func TestFoldV08_ToGraphQL_ClosedDelivered(t *testing.T) {
+	t0 := mustParse(t, "2026-05-21T10:00:00Z")
+	t1 := t0.Add(time.Hour)
+	c := Contract{
+		ID:             "C-tographql-closed-del",
+		Statement:      "make it done",
+		OwnerSessionID: "S-closed-del-001",
+		Status:         "closed",
+		ClosedReason:   "delivered",
+		CreatedAt:      t0,
+		UpdatedAt:      t1,
+		LastEventAt:    t1,
+	}
+	gc := toGraphQL(c)
+	if gc.Status != "CLOSED" {
+		t.Errorf("Status = %q, want CLOSED", gc.Status)
+	}
+	if gc.ClosedReason == nil {
+		t.Fatalf("ClosedReason is nil, want DELIVERED")
+	}
+	if *gc.ClosedReason != "DELIVERED" {
+		t.Errorf("ClosedReason = %q, want DELIVERED", *gc.ClosedReason)
+	}
+}
 
-	if got, want := len(state), 2; got != want {
-		t.Errorf("fold produced %d contracts; want 2 (user contracts with same deliverable are not deduped)", got)
+// TestFoldV08_ToGraphQL_ClosedAbandoned verifies the full toGraphQL projection
+// for a closed contract with reason=abandoned.
+func TestFoldV08_ToGraphQL_ClosedAbandoned(t *testing.T) {
+	t0 := mustParse(t, "2026-05-21T10:00:00Z")
+	c := Contract{
+		ID:             "C-tographql-closed-abn",
+		Statement:      "gave up",
+		OwnerSessionID: "S-closed-abn-001",
+		Status:         "closed",
+		ClosedReason:   "abandoned",
+		CreatedAt:      t0,
+		UpdatedAt:      t0,
+		LastEventAt:    t0,
+	}
+	gc := toGraphQL(c)
+	if gc.Status != "CLOSED" {
+		t.Errorf("Status = %q, want CLOSED", gc.Status)
+	}
+	if gc.ClosedReason == nil {
+		t.Fatalf("ClosedReason is nil, want ABANDONED")
+	}
+	if *gc.ClosedReason != "ABANDONED" {
+		t.Errorf("ClosedReason = %q, want ABANDONED", *gc.ClosedReason)
+	}
+}
+
+// ---- ContractFilter.closedReasons ----------------------------------------
+
+// TestMatches_ClosedReasonDelivered verifies that filtering by
+// closedReasons:[DELIVERED] returns only CLOSED+DELIVERED contracts and
+// excludes CLOSED+ABANDONED and SIGNED (open) contracts.
+func TestMatches_ClosedReasonDelivered(t *testing.T) {
+	t0 := mustParse(t, "2026-05-21T10:00:00Z")
+
+	delivered := Contract{
+		ID: "C-delivered", Status: "closed", ClosedReason: "delivered",
+		CreatedAt: t0, UpdatedAt: t0, LastEventAt: t0,
+	}
+	abandoned := Contract{
+		ID: "C-abandoned", Status: "closed", ClosedReason: "abandoned",
+		CreatedAt: t0, UpdatedAt: t0, LastEventAt: t0,
+	}
+	open := Contract{
+		ID: "C-open", Status: "open", ClosedReason: "",
+		CreatedAt: t0, UpdatedAt: t0, LastEventAt: t0,
+	}
+
+	filter := &graphql.ContractFilter{
+		ClosedReasons: []graphql.ContractReason{graphql.ContractReasonDelivered},
+	}
+
+	if !matches(delivered, filter) {
+		t.Error("delivered contract should match closedReasons:[DELIVERED] filter")
+	}
+	if matches(abandoned, filter) {
+		t.Error("abandoned contract should not match closedReasons:[DELIVERED] filter")
+	}
+	if matches(open, filter) {
+		t.Error("open contract should not match closedReasons:[DELIVERED] filter")
+	}
+}
+
+// TestMatches_ClosedReasonAbandoned verifies that filtering by
+// closedReasons:[ABANDONED] returns only CLOSED+ABANDONED contracts.
+func TestMatches_ClosedReasonAbandoned(t *testing.T) {
+	t0 := mustParse(t, "2026-05-21T10:00:00Z")
+
+	delivered := Contract{
+		ID: "C-del2", Status: "closed", ClosedReason: "delivered",
+		CreatedAt: t0, UpdatedAt: t0, LastEventAt: t0,
+	}
+	abandoned := Contract{
+		ID: "C-abn2", Status: "closed", ClosedReason: "abandoned",
+		CreatedAt: t0, UpdatedAt: t0, LastEventAt: t0,
+	}
+
+	filter := &graphql.ContractFilter{
+		ClosedReasons: []graphql.ContractReason{graphql.ContractReasonAbandoned},
+	}
+
+	if !matches(abandoned, filter) {
+		t.Error("abandoned contract should match closedReasons:[ABANDONED] filter")
+	}
+	if matches(delivered, filter) {
+		t.Error("delivered contract should not match closedReasons:[ABANDONED] filter")
+	}
+}
+
+// TestMatches_ClosedReasonBoth verifies that filtering by
+// closedReasons:[DELIVERED, ABANDONED] returns all closed contracts
+// regardless of reason, but excludes open contracts.
+func TestMatches_ClosedReasonBoth(t *testing.T) {
+	t0 := mustParse(t, "2026-05-21T10:00:00Z")
+
+	delivered := Contract{
+		ID: "C-del3", Status: "closed", ClosedReason: "delivered",
+		CreatedAt: t0, UpdatedAt: t0, LastEventAt: t0,
+	}
+	abandoned := Contract{
+		ID: "C-abn3", Status: "closed", ClosedReason: "abandoned",
+		CreatedAt: t0, UpdatedAt: t0, LastEventAt: t0,
+	}
+	open := Contract{
+		ID: "C-open3", Status: "open",
+		CreatedAt: t0, UpdatedAt: t0, LastEventAt: t0,
+	}
+
+	filter := &graphql.ContractFilter{
+		ClosedReasons: []graphql.ContractReason{
+			graphql.ContractReasonDelivered,
+			graphql.ContractReasonAbandoned,
+		},
+	}
+
+	if !matches(delivered, filter) {
+		t.Error("delivered contract should match closedReasons:[DELIVERED,ABANDONED] filter")
+	}
+	if !matches(abandoned, filter) {
+		t.Error("abandoned contract should match closedReasons:[DELIVERED,ABANDONED] filter")
+	}
+	if matches(open, filter) {
+		t.Error("open contract should not match closedReasons:[DELIVERED,ABANDONED] filter")
 	}
 }
